Add Lynx Rev1 ESP32 S3 support (#1271)
* Added FujiNet Lynx Rev1 ESP32-S3 PlatformIO build support in
platformio-sample.ini and new
build-platforms/platformio-fujinet-lynx-rev1.ini
* Added Lynx Rev1 web UI config at
data/webui/config/fujinet-lynx-rev1.yaml.
* Added new PINMAP_LYNX_S3 pin mapping with SD, UART, button, LED, assignments
* Enable/fix printf debug output for S3 based Lynx
* Added Lynx Rev1 hardware version detection/reporting
* Updated GPIO read/write helpers to use ESP-IDF gpio_get_level /
gpio_set_level since S3 has more GPIO than older ESP32
* Increased card-detect task stack size so it doesn't crash
* Added Lynx S3 LED initialization for BUS, BT, and WIFI LEDs
* Remove bogus COMLYNX reset interrupt setup
* Add to workflow build

+++ lib/device/comlynx/lynxFuji.cpp
@@ -273,7 +273,7 @@ void lynxFuji::comlynx_process()
     
     // Get the entire payload from Lynx
     uint16_t len = comlynx_recv_length();
-    Debug_printf("lynxFuji::comlynx_process - len: %ld, ", len);
+    Debug_printf("lynxFuji::comlynx_process - len: %ld, ", (long int)len);
 
     comlynx_recv_buffer(recvbuffer, len);
     if (comlynx_recv_ck()) {
